+++ packages/wise/src/wise/actions/wise_stack.py
@@ -11,5 +11,5 @@ Usage: wise stack FILES_OR_FILE_LIST [-o OUTPUT_FITS]
 Additional options:
 --output FILENAME, -o FILENAME: output file name (default=stack_img.fits)
 --nsigma NSIGMA, -n NSIGMA: clip background below NSIGMA level (default=0)
---nsigma_connected, -c: Keep only the brightest isolated structure
+--keep_brightest_only, -c: Keep only the brightest isolated structure
 '''